Phase 7.1d: Migrate TruncatedText to plain CSS
- Created plain CSS implementation with active/inactive states
- Converted touch device query media queries
- Maintained backward compatibility with .legacy.tsx wrapper
- Used CSS variables for link colors
