+++ src/app/content/highlights/components/TruncatedText.css
@@ -1,0 +1,63 @@
+/**
+ * TruncatedText Styles - Plain CSS version
+ */
+
+.truncated-text-link {
+  /* textStyle base */
+  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
+  /* linkStyle */
+  color: var(--link-color, #027EB5);
+  text-decoration: none;
+  text-align: left;
+  cursor: pointer;
+  border: none;
+  padding: 0;
+  margin: 0.4rem 0 0 0; /* cardPadding / 2 = 0.8 / 2 = 0.4 */
+  background: none;
+  font-size: 1.2rem;
+  line-height: 1.6rem;
+}
+
+.truncated-text-link:hover {
+  color: var(--link-hover-color, #0064A0);
+}
+
+.truncated-text-note {
+  /* textStyle base */
+  font-family: 'Helvetica Neue', Helvetica, Arial, sans-serif;
+  overflow: hidden;
+  white-space: pre-wrap;
+  word-break: break-word;
+  overflow-wrap: break-word;
+  font-size: 1.4rem;
+  line-height: 1.8rem;
+  margin: 0 0.4rem 0 0; /* cardPadding / 2 = 0.4 */
+  padding: 0;
+}
+
+.truncated-text-note.inactive {
+  text-overflow: ellipsis;
+  overflow: hidden;
+  max-height: 5.4rem; /* lineHeight * 3 = 1.8 * 3 = 5.4 */
+  display: -webkit-box;
+  -webkit-box-orient: vertical;
+  -webkit-line-clamp: 3;
+}
+
+.truncated-text-note.active + .truncated-text-link {
+  display: none;
+}
+
+/**
+ * Touch device query: (max-width: 75em) and (not all and (pointer: fine), (hover: none))
+ * This combines mobile breakpoint with touch device detection
+ */
+@media screen and (max-width: 75em) {
+  @media not all and (pointer: fine), (hover: none) {
+    .truncated-text-note {
+      padding: 0.8rem 1.6rem; /* cardPadding, cardPadding * 2 */
+      height: 15.2rem;
+      overflow: auto;
+    }
+  }
+}
